// src/interfaces/player.interface.ts
export interface Player{
    id: number;
    name: string;
    position: string;
    country: string;
    team: string;
    goals: number; 
    assist: number;
    motm: number;
    yellow_card: number;
    red_card: number;
    salary: number;
    overall: number;
    player_value: number;
    photo: string;

}